Added forward kinematics to get projected position of the right arm
Started using Canvas to draw the arm, but this proved to be a lot more finicky to play with. Reverted to using another ViewBox with the projected image. This worked better because the coordinates used are already scaled for this View object, which means no scaling and no additional display configurations to show the projected coordinates and drawings.